+++ back/src/modules/balanceModule/balance.type.ts
@@ -1,0 +1,11 @@
+export interface balance {
+        name: string;
+        balance: string;
+        symbol: string;
+        price?: any;
+}
+
+export interface balanceResponse {
+        status: number;
+        balances: balance[];
+}
