+++ back/src/modules/coinModule/coin.module.ts
@@ -5,10 +5,11 @@ import { Coin } from "../../models/coin.model";
 import { CoinController } from "./coin.controller";
 import { CoinService } from "./coin.service";
 import { Web3Module } from "../web3Module/web3.module";
+import { BalanceUtils } from "../balanceModule/balance.provider";
 
 @Module({
         imports: [SequelizeModule.forFeature([Coin]), Web3Module],
         controllers: [CoinController],
-        providers: [CoinService, TxService],
+        providers: [CoinService, TxService, BalanceUtils],
 })
 export class CoinModule {}
