+++ back/src/modules/loginModule/login.module.ts
@@ -1,9 +1,10 @@
 import { Module } from "@nestjs/common";
 import { LoginController } from "./login.controller";
 import { LoginService } from "./login.service";
+import { LoginProvider } from "./login.provider";
 
 @Module({
         controllers: [LoginController],
-        providers: [LoginService],
+        providers: [LoginService, LoginProvider],
 })
 export class LoginModule {}
