+++ back/src/modules/tradingModule/trading.type.ts
