+++ memo/2312/1231 - [Modify] DI.md
@@ -1,0 +1,17 @@
+현재 백엔드 구성을 살펴보면
+
+각 모듈에 모듈, 컨트롤러, 서비스가 있는데, 여기에 더해 utils 파일이 있고,
+
+여기서 함수를 직접 가져와서 사용한다.
+
+그런데 이건 nestJS의 핵심원칙인 DI를 위반하는 행동이다.
+
+이럴 경우 나중에 문제가 생길 수 있다는 피드백을 받게되었다.
+
+따라서 이를 수정해보도록 하자.
+
+Injectable class를 하나 만들고,
+
+여기에 utils 등의 다른 파일들에 있는 함수들이나 속성들을 옮겨주면 될 것이다.
+
+커스텀 타입이나 인터페이스는 .type.ts 파일에 모으고, import해주면 될 듯
